+++ CHANGES.md
@@ -3,6 +3,8 @@
 * update `Requires PHP` to 7.4 for `class-parser.php`
 * update `Requires WP` to 5.9
 * update `freemius/wordpress-sdk`
+* update `printf()` in `class Branches`
+* fix old `git-updater-pro` and `git-updater-additions` textdomains
 
 #### 12.4.0 / 2024-03-04
 * update `freemius/wordpress-sdk`